refactor(midnight-js): split indexer-public-data-provider monolith into 7 layered files
Phase 1 of #808 — pure structural extraction with no behavior change.
Splits the 902-LOC `indexer-public-data-provider.ts` into seven layered
modules to establish clear seams along which Phase 2 (lifecycle, config
object, class form) and the eventual Effect migration (#843) can land
without rewriting the package shape:

- `config.ts` — URL/protocol validation, extracted from the inline checks
- `transport.ts` — Apollo client construction (HttpLink + RetryLink + WS)
- `codec.ts` — pure decoders/transformers (parseHex*, to*, correlateDeployTxId)
- `mapping.ts` — codegen→domain mappers and shared type guards
- `observables.ts` — all RxJS pipelines and Apollo query/sub helpers
- `provider.ts` — inner factory returning a `PublicDataProvider`
- `index.ts` — public factory wrapper (still holds 9× assertIsContractAddress,
   removed in Phase 3)

Public surface preserved: `indexerPublicDataProvider(queryURL, subscriptionURL, webSocket?)`
keeps its positional signature; all previously-exported pure helpers
remain reachable from the package root. The one external call site
(`testkit-js/.../contract/providers.ts`) compiles unchanged.

Test imports updated from `'../indexer-public-data-provider'` to `'..'`
(import-path-only changes; assertions and behavior unchanged). All 64
unit tests pass; lint clean; build green; testkit-js downstream build
verified.

Out of scope (covered by subsequent PRs per spec):
- Phase 2: IndexerProviderConfig object, IndexerPublicDataProvider class,
  dispose() lifecycle, packages/types `dispose?()` (closes #820)
- Phase 3: wrapper consolidation + assertion hygiene
- Phase 4: polling → subscriptions
- Phase 5: generic block-offset observable

Spec: docs/specs/2026-06-08-indexer-provider-phase-1-2-restructure-lifecycle.md
